# apps/api/openapi.json
{
  "components": {
    "schemas": {
      "AddWatchlistRequest": {
        "additionalProperties": false,
        "properties": {
          "symbol": {
            "description": "6 位 A 股代码，例如 600519",
            "pattern": "^\\d{6}$",
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "symbol"
        ],
        "title": "AddWatchlistRequest",
        "type": "object"
      },
...
      "ErrorBody": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "description": "错误信封。HTTP 状态由 ``AppError.status_code`` 唯一决定（不在路由里硬编码）。",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "Freshness": {
        "description": "行情新鲜度。禁止把 stale 当 fresh 展示（spec §3.2）。",
        "enum": [
          "fresh",
          "stale"
        ],
        "title": "Freshness",
        "type": "string"
      },
...
      "ItemResponse_WatchlistAddedDTO_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/WatchlistAddedDTO"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "data",
          "request_id"
        ],
        "title": "ItemResponse[WatchlistAddedDTO]",
        "type": "object"
      },
      "JobDTO": {
        "additionalProperties": false,
        "properties": {
          "completed_steps": {
            "title": "Completed Steps",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current_step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Step"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "job_type": {
            "$ref": "#/components/schemas/JobType"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol"
          },
          "total_steps": {
            "title": "Total Steps",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "warnings": {
            "description": "非致命告警，例如分钟线不可得（spec §7.1）",
            "items": {},
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "id",
          "job_type",
          "status",
          "completed_steps",
          "total_steps",
          "created_at",
          "updated_at"
        ],
        "title": "JobDTO",
        "type": "object"
      },
      "JobStatus": {
        "enum": [
          "queued",
          "running",
          "succeeded",
          "failed"
        ],
        "title": "JobStatus",
        "type": "string"
      },
      "JobType": {
        "enum": [
          "instrument_backfill",
          "analysis_refresh",
          "quote_refresh",
          "prediction",
          "settlement"
        ],
        "title": "JobType",
        "type": "string"
      },
...
      "ListResponse_WatchlistItemDTO_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/WatchlistItemDTO"
            },
            "title": "Data",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "data",
          "page",
          "request_id"
        ],
        "title": "ListResponse[WatchlistItemDTO]",
        "type": "object"
      },
      "MarketDTO": {
        "additionalProperties": false,
        "description": "由 API 判定的市场时段；前端不读取本机时钟自行推断。",
        "properties": {
          "is_trading_day": {
            "title": "Is Trading Day",
            "type": "boolean"
          },
          "latest_trading_day": {
            "title": "Latest Trading Day",
            "type": "string"
          },
          "phase": {
            "$ref": "#/components/schemas/MarketPhase"
          }
        },
        "required": [
          "phase",
          "is_trading_day",
          "latest_trading_day"
        ],
        "title": "MarketDTO",
        "type": "object"
      },
      "MarketPhase": {
        "enum": [
          "closed",
          "pre_open",
          "call_auction",
          "morning",
          "lunch_break",
          "afternoon"
        ],
        "title": "MarketPhase",
        "type": "string"
      },
...
      "PageInfo": {
        "additionalProperties": false,
        "properties": {
          "has_more": {
            "default": false,
            "title": "Has More",
            "type": "boolean"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "title": "PageInfo",
        "type": "object"
      },
...
      "QuoteDTO": {
        "additionalProperties": false,
        "description": "近实时行情。\n\nspec §7.2 的示例只列出 price/change_percent/observed_at/source/freshness；\n这里额外带上 F2 明确要求的**成交量与量比**，以及 stale 时必须出现的 ``age_seconds``。",
        "properties": {
          "age_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "行情已过期多少秒。仅在 freshness=stale 时出现（spec §7）；fresh 时为 null",
            "title": "Age Seconds"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "change_percent": {
            "description": "price / previous_close - 1",
            "title": "Change Percent",
            "type": "number"
          },
          "fetched_at": {
            "description": "本系统取得该行情的时间（Asia/Shanghai）",
            "format": "date-time",
            "title": "Fetched At",
            "type": "string"
          },
          "freshness": {
            "$ref": "#/components/schemas/Freshness"
          },
          "high": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "High"
          },
          "low": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Low"
          },
          "market_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "上游明确提供的行情时间；当前来源未提供时为 null",
            "title": "Market Time"
          },
          "observed_at": {
            "description": "数据源观测时间（Asia/Shanghai）",
            "format": "date-time",
            "title": "Observed At",
            "type": "string"
          },
          "open": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Open"
          },
          "previous_close": {
            "title": "Previous Close",
            "type": "number"
          },
          "price": {
            "title": "Price",
            "type": "number"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "volume": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume"
          },
          "volume_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "量比（F2）",
            "title": "Volume Ratio"
          }
        },
        "required": [
          "symbol",
          "price",
          "previous_close",
          "change_percent",
          "observed_at",
          "fetched_at",
          "source",
          "freshness"
        ],
        "title": "QuoteDTO",
        "type": "object"
      },
...
      "ReorderWatchlistRequest": {
        "additionalProperties": false,
        "properties": {
          "symbols": {
            "description": "必须是当前自选股的一个全排列；缺项/多项/重复一律 400 INVALID_ARGUMENT",
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Symbols",
            "type": "array"
          }
        },
        "required": [
          "symbols"
        ],
        "title": "ReorderWatchlistRequest",
        "type": "object"
      },
...
      "WatchlistAddedDTO": {
        "additionalProperties": false,
        "description": "``POST /watchlist`` 的 data 体（spec §7.1）。",
        "properties": {
          "backfill_job": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobDTO"
              },
              {
                "type": "null"
              }
            ],
            "description": "首次添加时返回回补作业（HTTP 202）；已完成过回补则为 null（HTTP 201）"
          },
          "watchlist_item": {
            "$ref": "#/components/schemas/WatchlistItemDTO"
          }
        },
        "required": [
          "watchlist_item"
        ],
        "title": "WatchlistAddedDTO",
        "type": "object"
      },
      "WatchlistItemDTO": {
        "additionalProperties": false,
        "description": "自选股条目。\n\nspec §7.1 的 POST 示例只回 ``{\"symbol\", \"display_order\"}``；\n这里是它的超集：首页自选股表格需要 name / 行情 / 新鲜度，而\n**前端不得自行计算数据新鲜度**（spec §5.1），所以必须由 API 提供。",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_order": {
            "title": "Display Order",
            "type": "integer"
          },
          "is_current_universe_member": {
            "description": "false ⇒ 「已调出沪深300」：保留展示，停止新预测（spec §3.1）",
            "title": "Is Current Universe Member",
            "type": "boolean"
          },
          "market": {
            "$ref": "#/components/schemas/MarketDTO",
            "description": "API 判定的当前市场时段"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "quote": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QuoteDTO"
              },
              {
                "type": "null"
              }
            ],
            "description": "最新行情；从未取得行情时为 null（不编造默认值）"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "name",
          "display_order",
          "created_at",
          "is_current_universe_member",
          "market"
        ],
        "title": "WatchlistItemDTO",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "个人使用、仅沪深300成分股、OpenBB 统一数据网关、免费近实时数据、无自动交易。\n\n所有响应都带 request_id；列表使用游标分页（limit 默认 20、最大 100）。\n预测仅供研究，不构成投资建议。",
    "title": "A股 AI 研究助手 API",
    "version": "1.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/watchlist": {
      "get": {
        "operationId": "get_watchlist_api_v1_watchlist_get",
        "parameters": [
          {
            "description": "无填充 Base64URL 游标",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "无填充 Base64URL 游标",
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 80,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse_WatchlistItemDTO_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INVALID_ARGUMENT：代码、分页或时间参数无效"
          }
        },
        "summary": "自选股列表（含最新行情与新鲜度）",
        "tags": [
          "watchlist"
        ]
      },
      "post": {
        "operationId": "add_watchlist_item_api_v1_watchlist_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddWatchlistRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemResponse_WatchlistAddedDTO_"
                }
              }
            },
            "description": "此前已成功回补过该股票，无需新作业（backfill_job=null）"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemResponse_WatchlistAddedDTO_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INVALID_ARGUMENT：代码、分页或时间参数无效"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INSTRUMENT_NOT_FOUND：证券不存在或不在 MVP 市场范围"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "NOT_CURRENT_UNIVERSE_MEMBER / DUPLICATE_WATCHLIST_ITEM"
          }
        },
        "summary": "添加自选股（首次添加返回 202 + 回补任务）",
        "tags": [
          "watchlist"
        ]
      }
    },
    "/api/v1/watchlist/order": {
      "patch": {
        "operationId": "reorder_watchlist_api_v1_watchlist_order_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderWatchlistRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse_WatchlistItemDTO_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INVALID_ARGUMENT：代码、分页或时间参数无效"
          }
        },
        "summary": "重排自选股（symbols 必须是当前自选股的全排列）",
        "tags": [
          "watchlist"
        ]
      }
    },
    "/api/v1/watchlist/{symbol}": {
      "delete": {
        "operationId": "delete_watchlist_item_api_v1_watchlist__symbol__delete",
        "parameters": [
          {
            "description": "6 位 A 股代码，例如 600519",
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "description": "6 位 A 股代码，例如 600519",
              "examples": [
                "600519"
              ],
              "pattern": "^\\d{6}$",
              "title": "Symbol",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INVALID_ARGUMENT：代码、分页或时间参数无效"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INSTRUMENT_NOT_FOUND：证券不存在或不在 MVP 市场范围"
          }
        },
        "summary": "删除自选股",
        "tags": [
          "watchlist"
        ]
      }
    }
  }
}